backend/requirements.txt
fastapi
uvicorn[standard]
supabase==2.11.0
httpx==0.27.2
python-dotenv
pydantic
email-validator
sentry-sdk[fastapi]
google-genai
websockets>=14.0
pyngrok
